+++ src/test/java/org/openlmis/referencedata/service/GeographicZoneServiceTest.java
@@ -16,25 +16,48 @@
 package org.openlmis.referencedata.service;
 
 import static org.junit.Assert.assertEquals;
+import static org.mockito.Matchers.any;
+import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.when;
 
+import com.google.common.collect.Lists;
+import org.junit.Before;
 import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.mockito.InjectMocks;
 import org.mockito.Mock;
+import org.mockito.MockitoAnnotations;
 import org.mockito.runners.MockitoJUnitRunner;
+import org.openlmis.referencedata.domain.GeographicLevel;
 import org.openlmis.referencedata.domain.GeographicZone;
+import org.openlmis.referencedata.exception.ValidationMessageException;
+import org.openlmis.referencedata.repository.GeographicLevelRepository;
 import org.openlmis.referencedata.repository.GeographicZoneRepository;
+import org.openlmis.referencedata.util.Pagination;
+import org.springframework.data.domain.Page;
+import org.springframework.data.domain.Pageable;
 
 import java.util.Arrays;
 import java.util.Collection;
 import java.util.Collections;
+import java.util.HashMap;
+import java.util.List;
+import java.util.Map;
+import java.util.UUID;
 
 @RunWith(MockitoJUnitRunner.class)
 public class GeographicZoneServiceTest {
 
+  private static final String NAME = "name";
+  private static final String CODE = "code";
+  private static final String PARENT = "parent";
+  private static final String LEVEL_NUMBER = "levelNumber";
+
   @Mock
   private GeographicZoneRepository geographicZoneRepository;
+
+  @Mock
+  private GeographicLevelRepository geographicLevelRepository;
 
   @Mock
   private GeographicZone parent;
@@ -48,8 +71,29 @@ public class GeographicZoneServiceTest {
   @Mock
   private GeographicZone childOfChild;
 
+  @Mock
+  private GeographicLevel level;
+
+  @Mock
+  private Pageable pageable;
+
   @InjectMocks
   private GeographicZoneService geographicZoneService;
+
+  private UUID zoneId = UUID.randomUUID();
+  private List<GeographicZone> geographicZones;
+
+  @Before
+  public void setUp() {
+    MockitoAnnotations.initMocks(this);
+
+    when(child.getName()).thenReturn("zone-1");
+    when(secondChild.getName()).thenReturn("zone-2");
+    geographicZones = Lists.newArrayList(child, secondChild);
+
+    when(pageable.getPageSize()).thenReturn(10);
+    when(pageable.getPageNumber()).thenReturn(0);
+  }
 
   @Test
   public void shouldRetrieveOneDescendantWhenParentHasOneChild() {
@@ -95,5 +139,58 @@ public class GeographicZoneServiceTest {
         geographicZoneService.getAllZonesInHierarchy(parent);
 
     assertEquals(Collections.emptyList(), allZonesInHierarchy);
+  }
+
+  @Test(expected = ValidationMessageException.class)
+  public void shouldThrowExceptionIfThereNotProvidedForSearch() {
+    Map<String, Object> searchParams = new HashMap<>();
+    searchParams.put("some-param", "some-value");
+    geographicZoneService.search(searchParams, pageable);
+  }
+
+  @Test(expected = ValidationMessageException.class)
+  public void shouldThrowExceptionIfGeographicZoneDoesNotExist() {
+    when(geographicZoneRepository.findOne(any(UUID.class))).thenReturn(null);
+
+    Map<String, Object> searchParams = new HashMap<>();
+    searchParams.put(PARENT, "zone-code");
+    geographicZoneService.search(searchParams, pageable);
+  }
+
+  @Test(expected = ValidationMessageException.class)
+  public void shouldThrowExceptionIfProgramDoesNotExist() {
+    when(geographicLevelRepository.findByLevelNumber(any(Integer.class))).thenReturn(null);
+
+    Map<String, Object> searchParams = new HashMap<>();
+    searchParams.put(LEVEL_NUMBER, "1");
+    geographicZoneService.search(searchParams, pageable);
+  }
+
+  @Test
+  public void shouldReturnAllElementsIfNoSearchCriteriaProvided() {
+    when(geographicZoneRepository.findAll()).thenReturn(geographicZones);
+
+    Page<GeographicZone> actual = geographicZoneService.search(new HashMap<>(), pageable);
+    verify(geographicZoneRepository).findAll();
+    assertEquals(geographicZones, actual.getContent());
+  }
+
+  @Test
+  public void shouldSearchForRequisitionGroupsWithAllParametersProvided() {
+    when(geographicZoneRepository.findOne(any(UUID.class))).thenReturn(parent);
+    when(geographicLevelRepository.findByLevelNumber(any(Integer.class))).thenReturn(level);
+    when(geographicZoneRepository.search(any(String.class), any(String.class),
+        any(GeographicZone.class), any(GeographicLevel.class), any(Pageable.class)))
+        .thenReturn(Pagination.getPage(geographicZones, null, 2));
+
+    Map<String, Object> searchParams = new HashMap<>();
+    searchParams.put(NAME, "name");
+    searchParams.put(CODE, "code");
+    searchParams.put(PARENT, zoneId.toString());
+    searchParams.put(LEVEL_NUMBER, "1");
+
+    Page<GeographicZone> actual = geographicZoneService.search(searchParams, pageable);
+    verify(geographicZoneRepository).search("name", "code", parent, level, pageable);
+    assertEquals(geographicZones, actual.getContent());
   }
 }